+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Config to use sessionDurationSeconds and sessionRefreshThreshold instead of keepAliveSeconds. 
+  KeepAliveSeconds is now calculated from these two parameters and sessionDurationSeconds is used when starting a 
+  session.
 
 ### Added
+- Added getSessionInfo that logs sessionInfo. Only used for testing.
 
 ### Changed